+++ lib/runners/python.rb
@@ -6,6 +6,7 @@ module Runners
       capture3! "python", "--version"
       capture3! "pip", "--version"
       capture3! "pipenv", "--version"
+      super
     end
 
     # @see https://pip.pypa.io/en/stable/reference/pip_install