+++ lib/runners/ruby.rb
@@ -70,6 +70,7 @@ module Runners
       capture3! "ruby", "-v"
       capture3! "gem", "-v"
       capture3! "bundle", "-v"
+      super
     end
 
     def ruby_analyzer_command(*args)