+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Misc:
 - **Slim-Lint** Support Sass by default [#2297](https://github.com/sider/runners/pull/2297)
 - **RuboCop** Provide new recommended configuration [#2266](https://github.com/sider/runners/pull/2266)
 - **Metrics Complexity** Add workarounds for lizard's bug [#2249](https://github.com/sider/runners/pull/2249)
+- Show the tool's default version on setup [#2313](https://github.com/sider/runners/pull/2313)
 
 ## 0.47.0